pyproject.toml
[project]
name = "chapter4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>1.68",
    "langgraph==0.2.14",
    "pydantic-settings==2.4.0",
    "langchain>=0.2.15,<0.3.0",
    "pypdf==4.3.1",
    "langchain-community==0.2.13",
    "langchain-text-splitters==0.2.2",
]
[tool.uv]
dev-dependencies = [
    "ruff>=0.11.4",
    "ipykernel==6.29.5",
]

[tool.ruff]
select = [
    "F", # Flake8
    "B", # Black
    "I", # isort
    "E", # error
    "W"  # warning
]
ignore = []
fixable = ["ALL"]
exclude = [
    ".ruff_cache"
]
line-length = 119
indent-width = 4
target-version = "py312"
extend-include = ["*.ipynb"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
force-single-line = false
